requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

google-cloud-storage==2.18.2

python-multipart 

# Conda dependencies
numpy==1.26.4
opencv-python-headless==4.9.0.80
tensorflow==2.16.1
torch==2.2.2
torchvision==0.17.2
ultralytics==8.1.47

# Pip dependencies
werkzeug
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
idna==3.7
imageio==2.34.2
imbalanced-learn==0.12.3
importlib-metadata==7.1.0
importlib-resources==6.4.0
jinja2==3.1.3
joblib==1.4.0
keras==3.3.3
kiwisolver==1.4.5
lazy-loader==0.4
libclang==18.1.1
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.13.0
seaborn==0.13.2
segment-anything==1.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
thop==0.1.1-2209072238
threadpoolctl==3.4.0
tifffile==2024.7.2
tqdm==4.66.2
triton==2.2.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==3.0.2
wrapt==1.14.1
xgboost==2.0.3
zipp==3.18.1